Renamed the vertex and edge class
